refactor: step2 리뷰 반영
